+++ backend/src/main/java/com/spring/sad/album/domain/Album.java
@@ -2,6 +2,7 @@ package com.spring.sad.album.domain;
 
 import com.spring.sad.band.domain.Band;
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,8 +10,8 @@ import lombok.NoArgsConstructor;
 import java.util.ArrayList;
 import java.util.List;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@NoArgsConstructor
 @Entity
 public class Album {
     @Id
