+++ backend/src/main/java/com/spring/sad/album/repository/AlbumRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.album.domain;
+package com.spring.sad.album.repository;
 
+import com.spring.sad.album.domain.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AlbumRepository extends JpaRepository<Album, Long> {
