+++ backend/src/main/java/com/spring/sad/band/repository/BandRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.band.domain;
+package com.spring.sad.band.repository;
 
+import com.spring.sad.band.domain.Band;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BandRepository extends JpaRepository<Band, Long> {
