+++ backend/src/main/java/com/spring/sad/band/repository/CategoryRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.band.domain;
+package com.spring.sad.band.repository;
 
+import com.spring.sad.band.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
