+++ backend/src/main/java/com/spring/sad/image/repository/ImageRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.image.domain;
+package com.spring.sad.image.repository;
 
+import com.spring.sad.image.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
