+++ backend/src/main/java/com/spring/sad/member/repository/MemberBandRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.member.domain;
+package com.spring.sad.member.repository;
 
+import com.spring.sad.member.domain.MemberBand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberBandRepository extends JpaRepository<MemberBand, Long> {
