+++ backend/src/main/java/com/spring/sad/member/repository/ProfileRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.member.domain;
+package com.spring.sad.member.repository;
 
+import com.spring.sad.member.domain.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
