+++ backend/src/main/java/com/spring/sad/notification/repository/NotificationRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.notification.domain;
+package com.spring.sad.notification.repository;
 
+import com.spring.sad.notification.domain.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
