+++ backend/src/main/java/com/spring/sad/post/domain/PostImage.java
@@ -2,12 +2,13 @@ package com.spring.sad.post.domain;
 
 import com.spring.sad.image.domain.Image;
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@NoArgsConstructor
 @Entity
 public class PostImage {
     @Id
