+++ backend/src/main/java/com/spring/sad/post/domain/PostTag.java
@@ -1,6 +1,7 @@
 package com.spring.sad.post.domain;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,8 +9,8 @@ import lombok.NoArgsConstructor;
 import java.util.ArrayList;
 import java.util.List;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@NoArgsConstructor
 @Entity
 public class PostTag {
     @Id
@@ -17,11 +18,11 @@ public class PostTag {
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "postTags")
+    @JoinColumn(name = "post_id")
     private Post post;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "postTags")
+    @JoinColumn(name = "tag_id")
     private Tag tag;
 
     @Builder
