+++ backend/src/main/java/com/spring/sad/post/repository/PostImageRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.post.domain;
+package com.spring.sad.post.repository;
 
+import com.spring.sad.post.domain.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostImageRepository extends JpaRepository<PostImage, Long> {
