+++ backend/src/main/java/com/spring/sad/post/repository/PostTagRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.post.domain;
+package com.spring.sad.post.repository;
 
+import com.spring.sad.post.domain.PostTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostTagRepository extends JpaRepository<PostTag, Long> {
