+++ backend/src/main/java/com/spring/sad/post/repository/TagRepository.java
@@ -1,5 +1,6 @@
-package com.spring.sad.post.domain;
+package com.spring.sad.post.repository;
 
+import com.spring.sad.post.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
