+++ backend/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+spring:
+  config:
+    import: security/application-database.yml
